Server: added all request handlers.
Session: Added a method that wrappes the send General Respond logic
Mail: SetMailAsPacket gets the "from" user sepretly because we dont get it from the packet but from the session context.
Inbox: fillPacketWith mail splitted to another method that returns the Inbox from a user string, we need that logic sepretly.